# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Affordability": {
        "additionalProperties": false,
        "properties": {
          "band": {
            "enum": [
              "safe",
              "caution",
              "risk"
            ],
            "title": "Band",
            "type": "string"
          },
          "breakdown": {
            "$ref": "#/components/schemas/AffordabilityBreakdown"
          },
          "maxMonthlyHousingCostKRW": {
            "title": "Maxmonthlyhousingcostkrw",
            "type": "integer"
          },
          "rationale": {
            "items": {
              "type": "string"
            },
            "title": "Rationale",
            "type": "array"
          },
          "recommendedMonthlyHousingCostKRW": {
            "title": "Recommendedmonthlyhousingcostkrw",
            "type": "integer"
          }
        },
        "required": [
          "maxMonthlyHousingCostKRW",
          "recommendedMonthlyHousingCostKRW",
          "band",
          "breakdown",
          "rationale"
        ],
        "title": "Affordability",
        "type": "object"
      },
      "AffordabilityBreakdown": {
        "additionalProperties": false,
        "properties": {
          "buffer": {
            "title": "Buffer",
            "type": "integer"
          },
          "existingDebt": {
            "title": "Existingdebt",
            "type": "integer"
          },
          "livingCost": {
            "title": "Livingcost",
            "type": "integer"
          },
          "netIncome": {
            "title": "Netincome",
            "type": "integer"
          }
        },
        "required": [
          "netIncome",
          "livingCost",
          "existingDebt",
          "buffer"
        ],
        "title": "AffordabilityBreakdown",
        "type": "object"
      },
      "AnalyzeInternal": {
        "additionalProperties": false,
        "properties": {
          "dataFreshness": {
            "$ref": "#/components/schemas/InternalFreshness"
          },
          "ineligiblePolicies": {
            "items": {
              "$ref": "#/components/schemas/InternalIneligiblePolicy"
            },
            "title": "Ineligiblepolicies",
            "type": "array"
          },
          "ruleVersions": {
            "items": {
              "$ref": "#/components/schemas/InternalRuleVersion"
            },
            "title": "Ruleversions",
            "type": "array"
          }
        },
        "required": [
          "ruleVersions",
          "dataFreshness",
          "ineligiblePolicies"
        ],
        "title": "AnalyzeInternal",
        "type": "object"
      },
      "AnalyzeMeta": {
        "additionalProperties": false,
        "properties": {
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "engineVersion": {
            "title": "Engineversion",
            "type": "string"
          },
          "generatedAt": {
            "title": "Generatedat",
            "type": "string"
          },
          "region": {
            "$ref": "#/components/schemas/AnalyzeMetaRegion"
          }
        },
        "required": [
          "generatedAt",
          "engineVersion",
          "region",
          "disclaimer"
        ],
        "title": "AnalyzeMeta",
        "type": "object"
      },
      "AnalyzeMetaRegion": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "AnalyzeMetaRegion",
        "type": "object"
      },
      "AnalyzeResponse": {
        "additionalProperties": false,
        "properties": {
          "affordability": {
            "$ref": "#/components/schemas/Affordability"
          },
          "dataGrade": {
            "$ref": "#/components/schemas/DataGrade"
          },
          "internal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyzeInternal"
              },
              {
                "type": "null"
              }
            ],
            "description": "규칙 버전 · 신선도 · 부적격 상세 (SPEC 6.1). **인증된 요청에만 실린다.** 익명 응답에는 이 키가 존재하지 않는다 — `null` 과 부재는 다른 사실이다. `x-requires-role` 의 값은 이 필드를 받는 **가장 낮은 역할**이며, 상위 역할(rule_manager)도 함께 받는다.",
            "x-requires-role": "counselor"
          },
          "meta": {
            "$ref": "#/components/schemas/AnalyzeMeta"
          },
          "policies": {
            "items": {
              "$ref": "#/components/schemas/Policy"
            },
            "title": "Policies",
            "type": "array"
          },
          "provenance": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceItem"
            },
            "title": "Provenance",
            "type": "array"
          },
          "risk": {
            "$ref": "#/components/schemas/Risk"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/Scenario"
            },
            "title": "Scenarios",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "affordability",
          "scenarios",
          "policies",
          "risk",
          "summary",
          "meta",
          "provenance",
          "dataGrade"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
...
      "ChatMessage": {
        "properties": {
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "role": {
            "default": "user",
            "title": "Role",
            "type": "string"
          }
        },
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatRequest": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "History",
            "type": "array"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileRequest"
          }
        },
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "toolCalls": {
            "items": {
              "$ref": "#/components/schemas/ChatToolCall"
            },
            "title": "Toolcalls",
            "type": "array"
          }
        },
        "required": [
          "reply",
          "toolCalls",
          "mode",
          "provider"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ChatToolCall": {
        "additionalProperties": false,
        "properties": {
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "resultSummary": {
            "title": "Resultsummary",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "tool",
          "args",
          "resultSummary"
        ],
        "title": "ChatToolCall",
        "type": "object"
      },
...
      "DataGrade": {
        "additionalProperties": false,
        "description": "SPEC 2.4 — 등급값 + 사유 목록. 산정 대상은 `verification != our_choice` 다.",
        "properties": {
          "grade": {
            "anyOf": [
              {
                "enum": [
                  "A",
                  "B",
                  "C"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "`C` > `B` > `A` — 가장 나쁜 것이 이긴다 (SPEC 2.4). **`null` 은 「등급을 산정할 수 없다」이지 「문제 없음」이 아니다.** 2.4 의 `A` 는 「입력 전부 `verified` **이며** 신선도 기준 이내」라는 두 조건의 곱인데 신선도 임계가 아직 없어 뒤 절이 확정되지 않는다. 확정되지 않은 것을 참으로 읽어 `A` 를 내면 그것이 낙관적 등급이다. `null` 일 때 그 이유는 `reasons` 에 `freshness_not_evaluated` 로 **반드시** 실린다 — 사유 없는 `null` 은 소비자에게 「깨끗함」으로 읽힌다.",
            "title": "Grade"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/DataGradeReason"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "grade",
          "reasons"
        ],
        "title": "DataGrade",
        "type": "object"
      },
      "DataGradeReason": {
        "additionalProperties": false,
        "description": "등급 사유 하나. **원인 유형을 구분해 담는다** (SPEC 2.4).\n\n한 글자 등급으로 뭉치면 대응 주체의 구분이 사라진다 — `unverified`·`stale` 은 배치가,\n`pending_review` 는 규칙관리자가 푼다.",
        "properties": {
          "fact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "provenanceIndex": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenanceindex"
          },
          "type": {
            "enum": [
              "unverified",
              "stale",
              "pending_review",
              "freshness_not_evaluated"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "provenanceIndex",
          "fact",
          "message"
        ],
        "title": "DataGradeReason",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "모든 4xx·5xx 가 이 형태다 (SPEC 8.1 — 오류 봉투도 계약이다).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "InternalFreshness": {
        "additionalProperties": false,
        "description": "판정에 쓰인 지역 시세의 계보 (SPEC 6.1 「신선도」).\n\n`Region` 의 **레코드 단위 요약**이며 필드별 계보의 최악값이다 (SPEC 2.4).\n공개 응답의 `regions[].source` 는 화면 문구용 한 줄이라 `verification` 이 없다.",
        "properties": {
          "fetchedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetchedat"
          },
          "observedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "regionCode": {
            "title": "Regioncode",
            "type": "string"
          },
          "verification": {
            "title": "Verification",
            "type": "string"
          }
        },
        "required": [
          "regionCode",
          "verification",
          "observedAt",
          "fetchedAt"
        ],
        "title": "InternalFreshness",
        "type": "object"
      },
      "InternalIneligiblePolicy": {
        "additionalProperties": false,
        "description": "부적격·조건부 정책의 **문턱** (SPEC 6.1 「부적격 상세」).\n\n★ `reasons` 를 다시 싣지 않는다 (코디네이터 결정 Q6). 그것은 이미 공개 응답의\n  `policies[].reasons` 에 있다. 내부 정보의 값은 [왜 안 됐는지] 가 아니라\n  **[문턱이 얼마인지]** 다 — 상담원이 고객 앞에서 [연소득 때문에 안 됩니다] 가 아니라\n  [3,000만원이 문턱인데 4,200만원이십니다] 를 말할 수 있게 하는 것이 6.1 이 상담원에게\n  내부 정보를 주는 이유다.\n\n`criteria` 를 열거하지 않고 불투명 객체로 두는 이유는 정본이 하나여야 하기 때문이다 —\n그 모양은 `contracts/rule_draft.schema.json` 과 규칙 payload 가 정하고, 여기서 다시\n적으면 세 번째 정본이 생긴다.",
        "properties": {
          "criteria": {
            "additionalProperties": true,
            "title": "Criteria",
            "type": "object"
          },
          "policyId": {
            "title": "Policyid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "policyId",
          "status",
          "criteria"
        ],
        "title": "InternalIneligiblePolicy",
        "type": "object"
      },
...
      "Policy": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "maxAmountKRW": {
            "title": "Maxamountkrw",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rateRangePct": {
            "items": {
              "type": "number"
            },
            "title": "Raterangepct",
            "type": "array"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "eligible",
              "conditional",
              "ineligible"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "status",
          "reasons",
          "maxAmountKRW",
          "rateRangePct",
          "source",
          "disclaimer"
        ],
        "title": "Policy",
        "type": "object"
      },
...
      "ProfileRequest": {
        "properties": {
          "age": {
            "default": 28,
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Age",
            "type": "integer"
          },
          "annualIncomeKRW": {
            "default": 0,
            "maximum": 100000000000.0,
            "minimum": 0.0,
            "title": "Annualincomekrw",
            "type": "integer"
          },
          "existingDebtMonthlyKRW": {
            "default": 0,
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Existingdebtmonthlykrw",
            "type": "integer"
          },
          "householdSize": {
            "default": 1,
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Householdsize",
            "type": "integer"
          },
          "isHomeless": {
            "default": true,
            "title": "Ishomeless",
            "type": "boolean"
          },
          "isNewlywed": {
            "default": false,
            "title": "Isnewlywed",
            "type": "boolean"
          },
          "isSMEEmployee": {
            "default": false,
            "title": "Issmeemployee",
            "type": "boolean"
          },
          "liquidAssetsKRW": {
            "default": 0,
            "maximum": 100000000000.0,
            "minimum": 0.0,
            "title": "Liquidassetskrw",
            "type": "integer"
          },
          "monthlyNetIncomeKRW": {
            "default": 0,
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Monthlynetincomekrw",
            "type": "integer"
          },
          "preferredType": {
            "default": "any",
            "enum": [
              "jeonse",
              "monthly",
              "any"
            ],
            "title": "Preferredtype",
            "type": "string"
          },
          "regionCode": {
            "default": "",
            "title": "Regioncode",
            "type": "string"
          }
        },
        "title": "ProfileRequest",
        "type": "object"
      },
      "ProvenanceItem": {
        "additionalProperties": false,
        "description": "계보 배열의 한 항목 — **사실 하나**다 (D-13).\n\n`provenance` 를 통째로 중첩하는 이유는 계약 결정 #34 다. 생성물의 계보와 **같은\n모양**이어야 화면이 하나의 렌더러로 양쪽을 그린다. 평평하게 펴면 그 순간 응답의\n계보와 생성물의 계보가 다른 물건이 된다.",
        "properties": {
          "fact": {
            "description": "사람이 읽는 사실 이름. 기계 판독은 `factKind` 와 `provenance` 로 한다.",
            "title": "Fact",
            "type": "string"
          },
          "factKind": {
            "description": "데이터(D-5)와 모델 상수(Part 0-C)는 **같은 배열**에 들어간다 (D-13). 구분은 이 필드와 `provenance.source_kind` 가 진다.",
            "enum": [
              "region_field",
              "rule_version",
              "model_constant"
            ],
            "title": "Factkind",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceRecord"
          },
          "targets": {
            "description": "이 사실이 근거짓는 **응답 위치**의 JSON Pointer 목록 (RFC 6901 · SPEC 2.1). 좁게 특정할 수 없는 곳은 그 부분트리를 가리킨다.",
            "items": {
              "type": "string"
            },
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "fact",
          "factKind",
          "provenance",
          "targets"
        ],
        "title": "ProvenanceItem",
        "type": "object"
      },
...
      "Risk": {
        "additionalProperties": false,
        "properties": {
          "band": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Band",
            "type": "string"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/RiskFactor"
            },
            "title": "Factors",
            "type": "array"
          },
          "rationale": {
            "items": {
              "type": "string"
            },
            "title": "Rationale",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "score",
          "band",
          "factors",
          "rationale"
        ],
        "title": "Risk",
        "type": "object"
      },
...
      "Scenario": {
        "additionalProperties": false,
        "properties": {
          "components": {
            "$ref": "#/components/schemas/ScenarioComponents"
          },
          "depositKRW": {
            "title": "Depositkrw",
            "type": "integer"
          },
          "fitScore": {
            "title": "Fitscore",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "loanAmountKRW": {
            "title": "Loanamountkrw",
            "type": "integer"
          },
          "loanRatePct": {
            "title": "Loanratepct",
            "type": "number"
          },
          "monthlyEquivalentCostKRW": {
            "title": "Monthlyequivalentcostkrw",
            "type": "integer"
          },
          "monthlyRentKRW": {
            "title": "Monthlyrentkrw",
            "type": "integer"
          },
          "npv5yKRW": {
            "title": "Npv5Ykrw",
            "type": "integer"
          },
          "rationale": {
            "items": {
              "type": "string"
            },
            "title": "Rationale",
            "type": "array"
          },
          "schwabeIndexPct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "이 시나리오의 월환산비용 ÷ 월 실수령액(퍼센트 포인트). **월 실수령액이 0이면 이 비율은 정의되지 않으므로 `null` 이다.** 0.0 과 `null` 은 다른 사실이다 — 0.0 은 「주거비가 0이다」이고 `null` 은 「소득이 없어 잴 수 없다」다. 소비자는 `null` 을 0으로 대체하지 않는다.",
            "title": "Schwabeindexpct"
          },
          "tco5yKRW": {
            "title": "Tco5Ykrw",
            "type": "integer"
          },
          "type": {
            "enum": [
              "jeonse",
              "monthly"
            ],
            "title": "Type",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "affordable",
              "stretch",
              "unaffordable"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "type",
          "depositKRW",
          "monthlyRentKRW",
          "loanAmountKRW",
          "loanRatePct",
          "monthlyEquivalentCostKRW",
          "tco5yKRW",
          "npv5yKRW",
          "components",
          "fitScore",
          "schwabeIndexPct",
          "verdict",
          "rationale"
        ],
        "title": "Scenario",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "청년 주거 금융 의사결정 엔진 — 4대 결정론적 엔진 + LLM 상담 레이어",
    "title": "KB 첫집 나침반 API",
    "version": "2.0.0",
    "x-engine-version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analyze": {
      "post": {
        "description": "Main endpoint — runs E1~E4 and returns the full decision object.",
        "operationId": "analyze_endpoint_api_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "오류 봉투 (SPEC 8.1)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "오류 봉투 (SPEC 8.1)"
          }
        },
        "summary": "Analyze Endpoint"
      }
    },
...
    "/api/chat": {
      "post": {
        "description": "Natural-language consultation. Falls back to a template without an API key.",
        "operationId": "chat_endpoint_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse",
                  "additionalProperties": true,
                  "title": "Response Chat Endpoint Api Chat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "오류 봉투 (SPEC 8.1)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "오류 봉투 (SPEC 8.1)"
          }
        },
        "summary": "Chat Endpoint"
      }
    },
...
  },
...
}